COMMERCE-618 Rename directory to talend-installer
